+++ src/Shovel.js
@@ -1,8 +1,9 @@
 import {EventExchange} from './Exchange';
 
-export function Shovel(name, source, destination, cloneMessage) {
+export function Shovel(name, source, destination, options = {}) {
   const {broker: sourceBroker, exchange: sourceExchangeName, pattern, queue} = source;
   const {broker: destinationBroker, exchange: destinationExchangeName} = destination;
+  const {cloneMessage} = options;
 
   const sourceExchange = sourceBroker.getExchange(sourceExchangeName);
   if (!sourceExchange) {
@@ -14,7 +15,8 @@ export function Shovel(name, source, destination, cloneMessage) {
     throw new Error(`shovel ${name} destination exchange <${destinationExchangeName}> not found`);
   }
 
-  const consumerTag = `smq.shoveltag-${name}`;
+  const sameBroker = sourceBroker === destinationBroker;
+  const consumerTag = source.consumerTag || `smq.shoveltag-${name}`;
   const routingKeyPattern = pattern || '#';
   const events = EventExchange();
 
@@ -51,7 +53,9 @@ export function Shovel(name, source, destination, cloneMessage) {
 
   function onShovelMessage(routingKey, message) {
     const {content, properties} = messageHandler(message);
-    destinationExchange.publish(routingKey, content, {...properties, 'shovel-name': name});
+    const props = {...properties, 'source-exchange': sourceExchangeName};
+    if (!sameBroker) props['shovel-name'] = name;
+    destinationExchange.publish(routingKey, content, props);
     message.ack();
   }